refactor: apiserver error
