+++ apiserver/src/lib.rs
@@ -1,18 +1,14 @@
-use axum::{
-    http::StatusCode,
-    response::{IntoResponse, Response},
-};
 use nt_analyzer::Analyzer;
 use nt_cmd::config;
 use nt_io::PacketIO;
 use nt_modifier::Modifier;
 use nt_ruleset::expr_rule::ExprRuleset;
-use snafu::Snafu;
 use std::error::Error;
 use std::sync::Arc;
 use tokio::sync::{broadcast, RwLock};
 use tracing_subscriber::fmt::MakeWriter;
 
+pub mod error;
 pub mod file;
 pub mod save;
 pub mod service;
@@ -75,41 +71,4 @@ pub struct ServerConfig {
     pub engine_cancellation_token: tokio_util::sync::CancellationToken,
     pub program_cancellation_token: tokio_util::sync::CancellationToken,
     pub engine_handler: Option<tokio::task::JoinHandle<Result<(), Box<dyn Error + Send + Sync>>>>,
-}
-
-#[derive(Debug, Snafu)]
-pub enum ServiceError {
-    #[snafu(display("Failed to setup engine: {}", source))]
-    SetupEngine {
-        source: Box<dyn Error + Send + Sync>,
-    },
-
-    #[snafu(display("Failed to send shutdown signal"))]
-    ShutdownSignal,
-
-    #[snafu(display("{}", message))]
-    Common { message: String },
-
-    #[snafu(display("IO error {}", source))]
-    IoError { source: std::io::Error },
-
-    #[snafu(display("Serde error {}", source))]
-    SerdeError { source: serde_yaml::Error },
-}
-
-impl IntoResponse for ServiceError {
-    fn into_response(self) -> Response {
-        let (status, error_message) = match &self {
-            ServiceError::SetupEngine { .. } => {
-                (StatusCode::INTERNAL_SERVER_ERROR, self.to_string())
-            }
-            ServiceError::ShutdownSignal => (StatusCode::INTERNAL_SERVER_ERROR, self.to_string()),
-            ServiceError::Common { .. } => (StatusCode::INTERNAL_SERVER_ERROR, self.to_string()),
-            ServiceError::IoError { .. } => (StatusCode::INTERNAL_SERVER_ERROR, self.to_string()),
-            ServiceError::SerdeError { .. } => {
-                (StatusCode::INTERNAL_SERVER_ERROR, self.to_string())
-            }
-        };
-        (status, error_message).into_response()
-    }
 }
